// front/src/typechain/MusicShop.ts
export declare namespace MusicShop {
  export type AlbumStruct = {
    index: PromiseOrValue<BigNumberish>;
    uid: PromiseOrValue<string>;
    title: PromiseOrValue<string>;
    price: PromiseOrValue<BigNumberish>;
    quantity: PromiseOrValue<BigNumberish>;
  };

  export type AlbumStructOutput = [
    BigNumber,
    string,
    string,
    BigNumber,
    BigNumber
  ] & {
    index: BigNumber;
    uid: string;
    title: string;
    price: BigNumber;
    quantity: BigNumber;
  };
}